Move KeePass and KeePassHTTP into a keepass folder
